Split whole-run builders and projection helpers
