+++ apps/server/vibesensor/use_cases/run/post_analysis_executor.py
@@ -13,57 +13,38 @@ import aiosqlite
 from opentelemetry.trace import SpanKind
 
 from vibesensor.shared.ports import RunPersistence
-from vibesensor.shared.run_context_warning import (
-    RunContextWarningsInput,
-    normalize_run_context_warnings,
-)
 from vibesensor.shared.structured_logging import log_extra
 from vibesensor.shared.tracing import mark_span_error, start_span
-from vibesensor.shared.types.json_types import JsonObject, JsonValue, is_json_object
+from vibesensor.shared.types.json_types import JsonObject
 from vibesensor.shared.types.persisted_analysis import PersistedAnalysis
-from vibesensor.shared.types.raw_capture import RawCaptureManifest, RawRunCapture
+from vibesensor.shared.types.raw_capture import RawRunCapture
 from vibesensor.shared.types.run_schema import RunMetadata
 from vibesensor.shared.types.whole_run_analysis import WholeRunArtifactManifest
 from vibesensor.use_cases.diagnostics.orders.whole_run_contracts import OrderTraceSummary
 from vibesensor.use_cases.diagnostics.orders.whole_run_family_summaries import (
-    WHOLE_RUN_ORDER_FAMILY_SUMMARY_ARTIFACT_KEY,
     WholeRunOrderFamilySummaryArtifactBundle,
-    build_whole_run_order_family_summary_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.orders.whole_run_scoring import (
-    WHOLE_RUN_ORDER_TRACE_SUMMARY_ARTIFACT_KEY,
     WholeRunOrderTraceSummaryArtifactBundle,
-    build_whole_run_order_trace_summary_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.orders.whole_run_traces import (
-    WHOLE_RUN_ORDER_TRACE_ARTIFACT_KEY,
     WholeRunOrderTraceArtifactBundle,
-    build_whole_run_order_trace_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.spatial_evidence_contracts import (
     SpatialEvidenceSummary,
 )
 from vibesensor.use_cases.diagnostics.whole_run_context import (
-    WHOLE_RUN_CONTEXT_LABEL_ARTIFACT_KEY,
     WholeRunContextArtifactBundle,
-    build_whole_run_context_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.whole_run_diagnosis_contracts import (
     WholeRunDiagnosisSummary,
 )
-from vibesensor.use_cases.diagnostics.whole_run_diagnosis_ranking import (
-    build_whole_run_diagnosis_summaries,
-)
 from vibesensor.use_cases.diagnostics.whole_run_spatial_coherence import (
-    WHOLE_RUN_SPATIAL_COHERENCE_ARTIFACT_KEY,
     WholeRunSpatialCoherenceArtifactBundle,
-    build_whole_run_spatial_coherence_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.whole_run_spectra import (
     WholeRunSpectralArtifactBundle,
     WholeRunSpectralBuildResult,
-    WholeRunSpectralCoverageSummary,
-    build_whole_run_spectral_artifact_bundle,
 )
 from vibesensor.use_cases.diagnostics.whole_run_windows import WholeRunWindowPlan
 from vibesensor.use_cases.run.post_analysis_input import (
@@ -86,6 +67,33 @@ from vibesensor.use_cases.run.post_analysis_outcomes import (
     PostAnalysisExecutionRetryableFailure,
     PostAnalysisExecutionSuccess,
     is_retryable_post_analysis_error,
+)
+from vibesensor.use_cases.run.post_analysis_whole_run_builders import (
+    build_whole_run_artifacts,
+    build_whole_run_context_artifacts,
+    build_whole_run_order_family_summary_artifacts,
+    build_whole_run_order_trace_artifacts,
+    build_whole_run_order_trace_summary_artifacts,
+    build_whole_run_spatial_coherence_artifacts,
+    merge_whole_run_artifact_bundles,
+    whole_run_total_sample_count,
+)
+from vibesensor.use_cases.run.post_analysis_whole_run_projection import (
+    append_run_context_warnings,
+    append_whole_run_analysis_metadata,
+    append_whole_run_context,
+    append_whole_run_diagnosis_summaries,
+    append_whole_run_diagnosis_summary_metadata,
+    append_whole_run_order_family_summary_metadata,
+    append_whole_run_order_summaries,
+    append_whole_run_order_trace_metadata,
+    append_whole_run_order_trace_summary_metadata,
+    append_whole_run_spatial_coherence_metadata,
+    append_whole_run_spatial_summaries,
+    append_whole_run_spectral_metadata,
+    build_diagnosis_summary_rows,
+    ranked_whole_run_order_summaries,
+    ranked_whole_run_spatial_summaries,
 )
 
 LOGGER = logging.getLogger(__name__)
@@ -217,14 +225,6 @@ class WholeRunDiagnosisSummaryBuilder(Protocol):
         order_summaries: tuple[OrderTraceSummary, ...],
         spatial_summaries: tuple[SpatialEvidenceSummary, ...],
     ) -> tuple[WholeRunDiagnosisSummary, ...]: ...
-
-
-@dataclass(frozen=True, slots=True)
-class _StoredWholeRunArtifactBundle:
-    """Generic sidecar bundle shape for final manifest persistence."""
-
-    manifest: WholeRunArtifactManifest
-    artifact_contents: dict[str, bytes]
 
 
 type PostAnalysisStageStatus = Literal["ok", "skipped", "degraded", "failed"]
@@ -390,37 +390,37 @@ def resolve_whole_run_builders(
 ) -> ResolvedWholeRunBuilders:
     return ResolvedWholeRunBuilders(
         artifact_builder=(
-            _build_whole_run_artifacts
+            build_whole_run_artifacts
             if whole_run_artifact_builder is None
             else whole_run_artifact_builder
         ),
         context_builder=(
-            _build_whole_run_context_artifacts
+            build_whole_run_context_artifacts
             if whole_run_context_builder is None
             else whole_run_context_builder
         ),
         order_trace_builder=(
-            _build_whole_run_order_trace_artifacts
+            build_whole_run_order_trace_artifacts
             if whole_run_order_trace_builder is None
             else whole_run_order_trace_builder
         ),
         order_trace_summary_builder=(
-            _build_whole_run_order_trace_summary_artifacts
+            build_whole_run_order_trace_summary_artifacts
             if whole_run_order_trace_summary_builder is None
             else whole_run_order_trace_summary_builder
         ),
         order_family_summary_builder=(
-            _build_whole_run_order_family_summary_artifacts
+            build_whole_run_order_family_summary_artifacts
             if whole_run_order_family_summary_builder is None
             else whole_run_order_family_summary_builder
         ),
         spatial_coherence_builder=(
-            _build_whole_run_spatial_coherence_artifacts
+            build_whole_run_spatial_coherence_artifacts
             if whole_run_spatial_coherence_builder is None
             else whole_run_spatial_coherence_builder
         ),
         diagnosis_summary_builder=(
-            _build_whole_run_diagnosis_summaries
+            build_diagnosis_summary_rows
             if whole_run_diagnosis_summary_builder is None
             else whole_run_diagnosis_summary_builder
         ),
@@ -633,7 +633,7 @@ def run_whole_run_pipeline_stages(
             else:
                 context_bundle = builders.context_builder(
                     run=run_input,
-                    total_sample_count=_whole_run_total_sample_count(loaded.raw_capture_manifest),
+                    total_sample_count=whole_run_total_sample_count(loaded.raw_capture_manifest),
                 )
                 context_status = "degraded"
                 context_diagnostic = {"build_mode": "total_sample_count_fallback"}
@@ -822,7 +822,7 @@ def run_whole_run_pipeline_stages(
         )
 
     persist_artifacts_stage_start = time.monotonic()
-    merged_bundle = _merge_whole_run_artifact_bundles(
+    merged_bundle = merge_whole_run_artifact_bundles(
         spectral_bundle,
         context_bundle,
         order_trace_bundle,
@@ -914,14 +914,14 @@ def run_build_report_facts_stage(
     stored_artifact_manifest = whole_run_output.stored_artifact_manifest
 
     if spectral_result is not None:
-        summary = _append_whole_run_spectral_metadata(
+        summary = append_whole_run_spectral_metadata(
             summary,
             spectral_result.coverage_summary,
             spectral_bundle=spectral_bundle,
         )
-        summary = _append_run_context_warnings(summary, spectral_result.coverage_summary.warnings)
+        summary = append_run_context_warnings(summary, spectral_result.coverage_summary.warnings)
     if context_bundle is not None:
-        summary = _append_whole_run_context(summary, context_bundle)
+        summary = append_whole_run_context(summary, context_bundle)
 
     analysis_payload = summary.to_json_object()
     analysis_metadata_payload = analysis_payload.get("analysis_metadata")
@@ -930,21 +930,21 @@ def run_build_report_facts_stage(
     )
 
     if order_trace_bundle is not None:
-        summary = _append_whole_run_order_trace_metadata(summary, order_trace_bundle)
+        summary = append_whole_run_order_trace_metadata(summary, order_trace_bundle)
     if order_trace_summary_bundle is not None:
-        summary = _append_whole_run_order_trace_summary_metadata(
+        summary = append_whole_run_order_trace_summary_metadata(
             summary,
             order_trace_summary_bundle,
         )
     if order_family_summary_bundle is not None:
-        summary = _append_whole_run_order_summaries(summary, order_family_summary_bundle)
-        summary = _append_whole_run_order_family_summary_metadata(
+        summary = append_whole_run_order_summaries(summary, order_family_summary_bundle)
+        summary = append_whole_run_order_family_summary_metadata(
             summary,
             order_family_summary_bundle,
         )
     if spatial_coherence_bundle is not None:
-        summary = _append_whole_run_spatial_summaries(summary, spatial_coherence_bundle)
-        summary = _append_whole_run_spatial_coherence_metadata(
+        summary = append_whole_run_spatial_summaries(summary, spatial_coherence_bundle)
+        summary = append_whole_run_spatial_coherence_metadata(
             summary,
             spatial_coherence_bundle,
         )
@@ -952,23 +952,21 @@ def run_build_report_facts_stage(
         diagnosis_summaries = diagnosis_summary_builder(
             analysis_metadata=analysis_metadata,
             context_bundle=context_bundle,
-            order_summaries=_ranked_whole_run_order_summaries(
-                order_family_summary_bundle.summaries
-            ),
+            order_summaries=ranked_whole_run_order_summaries(order_family_summary_bundle.summaries),
             spatial_summaries=(
-                _ranked_whole_run_spatial_summaries(spatial_coherence_bundle.summaries)
+                ranked_whole_run_spatial_summaries(spatial_coherence_bundle.summaries)
                 if spatial_coherence_bundle is not None
                 else ()
             ),
         )
         if diagnosis_summaries:
-            summary = _append_whole_run_diagnosis_summaries(summary, diagnosis_summaries)
-            summary = _append_whole_run_diagnosis_summary_metadata(
+            summary = append_whole_run_diagnosis_summaries(summary, diagnosis_summaries)
+            summary = append_whole_run_diagnosis_summary_metadata(
                 summary,
                 diagnosis_summaries,
             )
     if stored_artifact_manifest is not None:
-        summary = _append_whole_run_analysis_metadata(summary, stored_artifact_manifest)
+        summary = append_whole_run_analysis_metadata(summary, stored_artifact_manifest)
 
     return summary, _make_stage_result(
         stage_name=stage_name,
@@ -1248,488 +1246,4 @@ def _persistence_failure_result(
         run_id=run_id,
         completed_error=completed_error,
         callback_errors=callback_errors,
-    )
-
-
-def _build_whole_run_artifacts(
-    *,
-    run_id: str,
-    metadata: RunMetadata,
-    raw_capture: RawRunCapture,
-) -> WholeRunSpectralBuildResult:
-    return build_whole_run_spectral_artifact_bundle(
-        run_id=run_id,
-        metadata=metadata,
-        raw_capture=raw_capture,
-    )
-
-
-def _build_whole_run_context_artifacts(
-    *,
-    run: PostAnalysisRunInput,
-    total_sample_count: int | None = None,
-    window_plan: WholeRunWindowPlan | None = None,
-) -> WholeRunContextArtifactBundle | None:
-    if total_sample_count is not None and total_sample_count < 0:
-        raise ValueError("whole-run context builder requires total_sample_count >= 0")
-    return build_whole_run_context_artifact_bundle(
-        run_id=run.run_id,
-        metadata=run.context,
-        samples=run.context_samples,
-        total_sample_count=total_sample_count,
-        window_plan=window_plan,
-    )
-
-
-def _whole_run_total_sample_count(manifest: RawCaptureManifest) -> int:
-    if manifest.sensors:
-        return max(int(sensor.sample_count) for sensor in manifest.sensors)
-    return max(0, int(manifest.total_samples))
-
-
-def _build_whole_run_order_trace_artifacts(
-    *,
-    run: PostAnalysisRunInput,
-    spectral_bundle: WholeRunSpectralArtifactBundle,
-    context_bundle: WholeRunContextArtifactBundle,
-) -> WholeRunOrderTraceArtifactBundle | None:
-    return build_whole_run_order_trace_artifact_bundle(
-        run_id=run.run_id,
-        metadata=run.context,
-        spectral_manifest=spectral_bundle.manifest,
-        spectral_artifact_contents=spectral_bundle.artifact_contents,
-        context_labels=context_bundle.labels,
-        samples=run.context_samples,
-        lang=run.language,
-    )
-
-
-def _build_whole_run_order_trace_summary_artifacts(
-    *,
-    order_trace_bundle: WholeRunOrderTraceArtifactBundle,
-    context_bundle: WholeRunContextArtifactBundle,
-) -> WholeRunOrderTraceSummaryArtifactBundle | None:
-    return build_whole_run_order_trace_summary_artifact_bundle(
-        order_trace_bundle=order_trace_bundle,
-        context_labels=context_bundle.labels,
-    )
-
-
-def _build_whole_run_order_family_summary_artifacts(
-    *,
-    order_trace_bundle: WholeRunOrderTraceArtifactBundle,
-    order_trace_summary_bundle: WholeRunOrderTraceSummaryArtifactBundle,
-    context_bundle: WholeRunContextArtifactBundle,
-) -> WholeRunOrderFamilySummaryArtifactBundle | None:
-    return build_whole_run_order_family_summary_artifact_bundle(
-        order_trace_bundle=order_trace_bundle,
-        order_trace_summary_bundle=order_trace_summary_bundle,
-        context_labels=context_bundle.labels,
-    )
-
-
-def _build_whole_run_spatial_coherence_artifacts(
-    *,
-    run: PostAnalysisRunInput,
-    spectral_bundle: WholeRunSpectralArtifactBundle,
-    context_bundle: WholeRunContextArtifactBundle,
-    order_trace_bundle: WholeRunOrderTraceArtifactBundle,
-) -> WholeRunSpatialCoherenceArtifactBundle | None:
-    if not order_trace_bundle.points:
-        return None
-    return build_whole_run_spatial_coherence_artifact_bundle(
-        order_trace_bundle=order_trace_bundle,
-        spectral_manifest=spectral_bundle.manifest,
-        spectral_artifact_contents=spectral_bundle.artifact_contents,
-        context_labels=context_bundle.labels,
-        samples=run.samples,
-        lang=run.language,
-    )
-
-
-def _merge_whole_run_artifact_bundles(
-    *bundles: (
-        WholeRunSpectralArtifactBundle
-        | WholeRunContextArtifactBundle
-        | WholeRunOrderTraceArtifactBundle
-        | WholeRunOrderTraceSummaryArtifactBundle
-        | WholeRunOrderFamilySummaryArtifactBundle
-        | WholeRunSpatialCoherenceArtifactBundle
-        | None
-    ),
-) -> _StoredWholeRunArtifactBundle | None:
-    active_bundles = [bundle for bundle in bundles if bundle is not None]
-    if not active_bundles:
-        return None
-    base_manifest = active_bundles[0].manifest
-    merged_artifacts = list(base_manifest.artifacts)
-    merged_contents = dict(active_bundles[0].artifact_contents)
-    for bundle in active_bundles[1:]:
-        manifest = bundle.manifest
-        if (
-            manifest.run_id != base_manifest.run_id
-            or manifest.relative_dir != base_manifest.relative_dir
-            or manifest.window_policy != base_manifest.window_policy
-            or manifest.total_window_count != base_manifest.total_window_count
-        ):
-            raise ValueError("whole-run artifact bundles must share the same run/window plan")
-        for artifact in manifest.artifacts:
-            if artifact.artifact_key in merged_contents:
-                raise ValueError(
-                    "whole-run artifact bundles must not reuse artifact keys: "
-                    f"{artifact.artifact_key}"
-                )
-            if artifact.artifact_key not in bundle.artifact_contents:
-                raise ValueError(
-                    f"whole-run artifact bundle missing bytes for {artifact.artifact_key}"
-                )
-            merged_artifacts.append(artifact)
-        merged_contents.update(bundle.artifact_contents)
-    return _StoredWholeRunArtifactBundle(
-        manifest=WholeRunArtifactManifest(
-            run_id=base_manifest.run_id,
-            relative_dir=base_manifest.relative_dir,
-            window_policy=base_manifest.window_policy,
-            total_window_count=base_manifest.total_window_count,
-            artifacts=tuple(merged_artifacts),
-            created_at=base_manifest.created_at,
-            schema_version=base_manifest.schema_version,
-            storage_type=base_manifest.storage_type,
-        ),
-        artifact_contents=merged_contents,
-    )
-
-
-def _append_whole_run_analysis_metadata(
-    summary: PersistedAnalysis,
-    manifest: WholeRunArtifactManifest,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    sensor_ids = sorted(
-        {artifact.sensor_id for artifact in manifest.artifacts if artifact.sensor_id is not None}
-    )
-    analysis_metadata["whole_run_artifacts_available"] = True
-    analysis_metadata["whole_run_window_count"] = int(manifest.total_window_count)
-    analysis_metadata["whole_run_sensor_count"] = len(sensor_ids)
-    analysis_metadata["whole_run_artifact_count"] = len(manifest.artifacts)
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_spectral_metadata(
-    summary: PersistedAnalysis,
-    coverage_summary: WholeRunSpectralCoverageSummary,
-    *,
-    spectral_bundle: WholeRunSpectralArtifactBundle | None,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_spectral_available"] = spectral_bundle is not None
-    analysis_metadata["whole_run_spectral_window_count"] = (
-        spectral_bundle.manifest.total_window_count if spectral_bundle is not None else 0
-    )
-    analysis_metadata["whole_run_spectral_sensor_window_count"] = (
-        coverage_summary.total_sensor_window_count
-    )
-    analysis_metadata["whole_run_spectral_full_sensor_window_count"] = (
-        coverage_summary.full_sensor_window_count
-    )
-    analysis_metadata["whole_run_spectral_partial_sensor_window_count"] = (
-        coverage_summary.partial_sensor_window_count
-    )
-    analysis_metadata["whole_run_spectral_missing_sensor_window_count"] = (
-        coverage_summary.missing_sensor_window_count
-    )
-    analysis_metadata["whole_run_spectral_empty_sensor_window_count"] = (
-        coverage_summary.empty_sensor_window_count
-    )
-    analysis_metadata["whole_run_spectral_gap_count"] = coverage_summary.gap_count
-    analysis_metadata["whole_run_spectral_overlap_count"] = coverage_summary.overlap_count
-    analysis_metadata["whole_run_spectral_dropped_chunk_count"] = (
-        coverage_summary.dropped_chunk_count
-    )
-    analysis_metadata["whole_run_spectral_late_packet_chunk_count"] = getattr(
-        coverage_summary, "late_packet_chunk_count", 0
-    )
-    analysis_metadata["whole_run_spectral_udp_ingest_queue_drop_count"] = getattr(
-        coverage_summary, "udp_ingest_queue_drop_count", 0
-    )
-    analysis_metadata["whole_run_spectral_queue_overflow_chunk_count"] = (
-        coverage_summary.queue_overflow_chunk_count
-    )
-    analysis_metadata["whole_run_spectral_invalid_chunk_count"] = (
-        coverage_summary.invalid_chunk_count
-    )
-    analysis_metadata["whole_run_spectral_write_error_chunk_count"] = (
-        coverage_summary.write_error_chunk_count
-    )
-    analysis_metadata["whole_run_spectral_sample_rate_mismatch_sensor_count"] = (
-        coverage_summary.sample_rate_mismatch_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_sample_rate_unverified_sensor_count"] = (
-        coverage_summary.sample_rate_unverified_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_unanchored_sensor_count"] = (
-        coverage_summary.unanchored_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_legacy_sensor_count"] = (
-        coverage_summary.legacy_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_sync_unverified_sensor_count"] = (
-        coverage_summary.sync_unverified_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_stale_sync_sensor_count"] = (
-        coverage_summary.stale_sync_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_high_rtt_sensor_count"] = (
-        coverage_summary.high_rtt_sensor_count
-    )
-    analysis_metadata["whole_run_spectral_coverage_confidence"] = (
-        coverage_summary.coverage_confidence
-    )
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_run_context_warnings(
-    summary: PersistedAnalysis,
-    warnings: RunContextWarningsInput,
-) -> PersistedAnalysis:
-    if not warnings:
-        return summary
-    payload = summary.to_json_object()
-    existing = payload.get("warnings")
-    warnings_payload: list[JsonValue] = []
-    if isinstance(existing, list):
-        warnings_payload.extend(item for item in existing if is_json_object(item))
-    for warning in normalize_run_context_warnings(warnings):
-        warning_payload: JsonObject = {
-            "code": warning.code,
-            "severity": warning.severity,
-            "applies_to": warning.applies_to,
-            "title": warning.title,
-            "detail": warning.detail,
-        }
-        warnings_payload.append(warning_payload)
-    payload["warnings"] = warnings_payload
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_context(
-    summary: PersistedAnalysis,
-    bundle: WholeRunContextArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    payload["whole_run_context_intervals"] = [
-        interval.to_json_object() for interval in bundle.intervals
-    ]
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_context_available"] = True
-    analysis_metadata["whole_run_context_window_count"] = int(bundle.manifest.total_window_count)
-    analysis_metadata["whole_run_context_interval_count"] = len(bundle.intervals)
-    analysis_metadata["whole_run_context_full_window_count"] = sum(
-        1 for label in bundle.labels if label.context_coverage == "full"
-    )
-    analysis_metadata["whole_run_context_partial_window_count"] = sum(
-        1 for label in bundle.labels if label.context_coverage == "partial"
-    )
-    analysis_metadata["whole_run_context_missing_window_count"] = sum(
-        1 for label in bundle.labels if label.context_coverage == "missing"
-    )
-    analysis_metadata["whole_run_context_missing_speed_window_count"] = sum(
-        1 for label in bundle.labels if label.speed_validity == "missing"
-    )
-    analysis_metadata["whole_run_context_missing_rpm_window_count"] = sum(
-        1 for label in bundle.labels if label.rpm_validity == "missing"
-    )
-    analysis_metadata["whole_run_context_stale_speed_window_count"] = sum(
-        1 for label in bundle.labels if label.speed_is_stale
-    )
-    analysis_metadata["whole_run_context_stale_rpm_window_count"] = sum(
-        1 for label in bundle.labels if label.rpm_is_stale
-    )
-    analysis_metadata["whole_run_context_labels_artifact_key"] = (
-        WHOLE_RUN_CONTEXT_LABEL_ARTIFACT_KEY
-    )
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_order_trace_metadata(
-    summary: PersistedAnalysis,
-    bundle: WholeRunOrderTraceArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_order_traces_available"] = True
-    analysis_metadata["whole_run_order_trace_point_count"] = len(bundle.points)
-    analysis_metadata["whole_run_order_trace_candidate_count"] = len(
-        {point.hypothesis_key for point in bundle.points}
-    )
-    analysis_metadata["whole_run_order_trace_artifact_key"] = WHOLE_RUN_ORDER_TRACE_ARTIFACT_KEY
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_order_trace_summary_metadata(
-    summary: PersistedAnalysis,
-    bundle: WholeRunOrderTraceSummaryArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_order_trace_summaries_available"] = True
-    analysis_metadata["whole_run_order_trace_summary_count"] = len(bundle.summaries)
-    analysis_metadata["whole_run_order_trace_summary_artifact_key"] = (
-        WHOLE_RUN_ORDER_TRACE_SUMMARY_ARTIFACT_KEY
-    )
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_order_family_summary_metadata(
-    summary: PersistedAnalysis,
-    bundle: WholeRunOrderFamilySummaryArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_order_family_summaries_available"] = True
-    analysis_metadata["whole_run_order_family_summary_count"] = len(bundle.summaries)
-    analysis_metadata["whole_run_order_family_summary_artifact_key"] = (
-        WHOLE_RUN_ORDER_FAMILY_SUMMARY_ARTIFACT_KEY
-    )
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_spatial_coherence_metadata(
-    summary: PersistedAnalysis,
-    bundle: WholeRunSpatialCoherenceArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_spatial_coherence_available"] = True
-    analysis_metadata["whole_run_spatial_coherence_window_count"] = len(bundle.windows)
-    analysis_metadata["whole_run_spatial_coherence_candidate_count"] = len(
-        {row.candidate_key for row in bundle.windows}
-    )
-    analysis_metadata["whole_run_spatial_coherence_summary_count"] = len(bundle.summaries)
-    analysis_metadata["whole_run_spatial_coherence_artifact_key"] = (
-        WHOLE_RUN_SPATIAL_COHERENCE_ARTIFACT_KEY
-    )
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_order_summaries(
-    summary: PersistedAnalysis,
-    bundle: WholeRunOrderFamilySummaryArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    payload["whole_run_order_summaries"] = [
-        row.to_json_object() for row in _ranked_whole_run_order_summaries(bundle.summaries)
-    ]
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_spatial_summaries(
-    summary: PersistedAnalysis,
-    bundle: WholeRunSpatialCoherenceArtifactBundle,
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    payload["whole_run_spatial_summaries"] = [
-        row.to_json_object() for row in _ranked_whole_run_spatial_summaries(bundle.summaries)
-    ]
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_diagnosis_summaries(
-    summary: PersistedAnalysis,
-    diagnosis_summaries: tuple[WholeRunDiagnosisSummary, ...],
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    payload["whole_run_diagnosis_summaries"] = [row.to_json_object() for row in diagnosis_summaries]
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _append_whole_run_diagnosis_summary_metadata(
-    summary: PersistedAnalysis,
-    diagnosis_summaries: tuple[WholeRunDiagnosisSummary, ...],
-) -> PersistedAnalysis:
-    payload = summary.to_json_object()
-    analysis_metadata = payload.get("analysis_metadata")
-    if not isinstance(analysis_metadata, dict):
-        analysis_metadata = {}
-        payload["analysis_metadata"] = analysis_metadata
-    analysis_metadata["whole_run_diagnosis_summaries_available"] = True
-    analysis_metadata["whole_run_diagnosis_summary_count"] = len(diagnosis_summaries)
-    return PersistedAnalysis.from_json_object(payload)
-
-
-def _build_whole_run_diagnosis_summaries(
-    *,
-    analysis_metadata: Mapping[str, object],
-    context_bundle: WholeRunContextArtifactBundle,
-    order_summaries: tuple[OrderTraceSummary, ...],
-    spatial_summaries: tuple[SpatialEvidenceSummary, ...],
-) -> tuple[WholeRunDiagnosisSummary, ...]:
-    return build_whole_run_diagnosis_summaries(
-        analysis_metadata=analysis_metadata,
-        context_intervals=context_bundle.intervals,
-        order_summaries=order_summaries,
-        spatial_summaries=spatial_summaries,
-    )
-
-
-def _ranked_whole_run_order_summaries(
-    summaries: tuple[OrderTraceSummary, ...],
-) -> tuple[OrderTraceSummary, ...]:
-    return tuple(
-        sorted(
-            summaries,
-            key=lambda summary: (
-                -summary.lock_score,
-                -summary.matched_window_count,
-                -summary.support_ratio,
-                -(summary.peak_intensity_db if summary.peak_intensity_db is not None else -1.0),
-                -summary.reference_coverage_ratio,
-                summary.hypothesis_key,
-            ),
-        )
-    )
-
-
-def _ranked_whole_run_spatial_summaries(
-    summaries: tuple[SpatialEvidenceSummary, ...],
-) -> tuple[SpatialEvidenceSummary, ...]:
-    return tuple(
-        sorted(
-            summaries,
-            key=lambda summary: (
-                -summary.coherent_window_count,
-                -summary.supporting_window_count,
-                -(summary.coherence_ratio if summary.coherence_ratio is not None else -1.0),
-                -(
-                    summary.location_separation_db
-                    if summary.location_separation_db is not None
-                    else -1.0
-                ),
-                -(summary.dominance_ratio if summary.dominance_ratio is not None else -1.0),
-                summary.candidate_key,
-            ),
-        )
     )
